LPD-61553 Move formatAccessibility and import it
