+++ modules/apps/layout/layout-js-components-web/test/__lib__/index.ts
@@ -4,3 +4,4 @@
  */
 
 export {default as checkAccessibility} from './checkAccessibility';
+export {default as formatAccessibility} from './formatAccessibility';
